requirements.txt
opencv-python
numpy
Pillow
torchvision
torch